bot/requirements.txt
aiogram==3.4.1
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
pydantic==2.5.3
